Fix a crash when disabling the status bar
A new custom status bar was previously added to the program, but the old
bar remained in the mainwindow.ui form. It was this bar, not the new
one, that was disabled in the menu, which caused the program to crash.

The ability to save the status bar state to the settings file has also
been added.